pyspider/requirements_additions.txt
Flask-RESTful==0.3.4
colander
newspaper
goose-extractor
dateutil